+++ packages/FormElement/src/components/Label/Label.styles.js
@@ -1,6 +1,7 @@
 import { css } from "styled-components";
 import tokens from "@paprika/tokens";
 import stylers from "@paprika/stylers";
+import { toInt } from "@paprika/stylers/lib/helpers";
 
 export const ruleStyles = css`
   color: ${tokens.textColor.subtle};
@@ -12,7 +13,10 @@ const labelStyles = css`
   display: inline-block;
   font-size: inherit;
   font-weight: bold;
-  margin: 0 ${stylers.spacer(4)} ${tokens.spaceSm} 0;
+  ${({ hasInstructionsShowing }) =>
+    `margin: 0 ${stylers.spacer(4)} ${
+      hasInstructionsShowing ? toInt(tokens.space) + toInt(tokens.spaceSm) : toInt(tokens.space) * 2
+    }px 0;`}
   padding: 0;
   position: relative;