+++ packages/Input/src/Input.js
@@ -28,8 +28,6 @@ const propTypes = {
   /** Displays an icon inside the input. */
   icon: PropTypes.node,
 
-  inputRef: PropTypes.oneOfType([PropTypes.func, PropTypes.shape({ current: PropTypes.instanceOf(Element) })]),
-
   /** If true it makes the input disabled. */
   isDisabled: PropTypes.bool,
 
@@ -60,7 +58,6 @@ const defaultProps = {
   hasClearButton: false,
   hasError: false,
   icon: null,
-  inputRef: () => {},
   isDisabled: false,
   isReadOnly: false,
   onChange: () => {},
@@ -70,7 +67,7 @@ const defaultProps = {
   value: null,
 };
 
-const Input = props => {
+const Input = React.forwardRef((props, ref) => {
   const inputClearHandler = e => {
     e.target.value = "";
     props.onChange(e);
@@ -102,9 +99,9 @@ const Input = props => {
 
   const {
     a11yText,
+    clearIcon,
     className,
     icon,
-    inputRef, // TODO: use useImperativeHandle()
     isDisabled,
     isReadOnly,
     hasClearButton,
@@ -146,13 +143,13 @@ const Input = props => {
         data-pka-anchor="input"
         disabled={isDisabled}
         readOnly={isReadOnly}
-        ref={inputRef}
+        ref={ref}
         {...moreProps}
       />
       {renderClear()}
     </div>
   );
-};
+});
 
 Input.displayName = "Input";
 Input.propTypes = propTypes;