+++ packages/Input/stories/examples/InputExample.js
@@ -1,10 +1,10 @@
 import React from "react";
 import Input from "../../src";
 
-const InputExample = props => {
+const InputExample = React.forwardRef((props, ref) => {
   const [value, setValue] = React.useState(props.value || "");
 
-  return <Input {...props} onChange={e => setValue(e.target.value)} value={value} />;
-};
+  return <Input {...props} onChange={e => setValue(e.target.value)} ref={ref} value={value} />;
+});
 
 export default InputExample;