+++ packages/SidePanel/src/components/Footer/Footer.styles.js
@@ -12,7 +12,7 @@ export const Footer = styled.div`
   box-sizing: border-box;
   display: flex;
   flex-shrink: 0;
-  height: ${props => (props.height ? props.height : stylers.spacer(9))};
+  height: ${props => (props.height ? `${props.height}px` : stylers.spacer(9))};
   padding: ${stylers.spacer(2)} ${stylers.spacer(3)};
   position: relative;
   transition: opacity 0.3s ease-in;