+++ packages/TimePicker/src/TimePicker.styles.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+import tokens from "@paprika/tokens";
+
+export const TimePicker = styled.div`
+  position: relative;
+
+  .timeinput-picker {
+    position: absolute;
+  }
+
+  input[type="text"][disabled] {
+    background: ${tokens.color.blackLighten70};
+    cursor: not-allowed;
+  }
+`;