+++ packages/ListBox/CHANGELOG.md
@@ -1,1 +1,18 @@
 # CHANGELOG
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [0.9.0] - 2020-06-16
+
+### Changed
+
+- Add onKeyUp event for `SPACE`, `ESC` and `ENTER` before onKeyDown was handling all key interaction now its only responsible for `↑`,`↓` keys.
+  this might break testing that depends onKeyDown event.
+
+## [0.9.3] - 2020-08-13
+
+### Changed
+
+- Add `label` prop on `ListBox.Trigger` component to support custom text as the label.