+++ packages/Toast/CHANGELOG.md
@@ -1,1 +1,7 @@
 # CHANGELOG
+
+## 2020-07-29
+
+### Added
+
+- Added imperative api with focus method. Set close button to semantic false [@tristanjasper](https://github.com/tristanjasper).